+++ experiment/errors_test.go
@@ -1,8 +1,6 @@
 package experiment
 
-import (
-	"testing"
-)
+import "testing"
 
 func TestErrNoRuleFieldsSpecifiedError(t *testing.T) {
 	want := "no rule fields specified"
@@ -33,33 +31,6 @@ func TestInvalidSortFieldErrorError(t *testing.T) {
 func TestInvalidRuleFieldErrorError(t *testing.T) {
 	e := InvalidRuleFieldError("bob")
 	want := "invalid rule field: bob"
-	got := e.Error()
-	if got != want {
-		t.Errorf("Error() got: %s, want: %s", got, want)
-	}
-}
-
-func TestInvalidAggregatorNameErrorError(t *testing.T) {
-	e := InvalidAggregatorNameError("bob")
-	want := "invalid aggregator name: bob"
-	got := e.Error()
-	if got != want {
-		t.Errorf("Error() got: %s, want: %s", got, want)
-	}
-}
-
-func TestAggregatorNameClashErrorError(t *testing.T) {
-	e := AggregatorNameClashError("bob")
-	want := "aggregator name clashes with field name: bob"
-	got := e.Error()
-	if got != want {
-		t.Errorf("Error() got: %s, want: %s", got, want)
-	}
-}
-
-func TestAggregatorNameReservedErrorError(t *testing.T) {
-	e := AggregatorNameReservedError("bob")
-	want := "aggregator name reserved: bob"
 	got := e.Error()
 	if got != want {
 		t.Errorf("Error() got: %s, want: %s", got, want)