+++ internal/misc_test.go
@@ -104,7 +104,7 @@ func TestGeneratePoints(t *testing.T) {
 	}
 }
 
-func TestStringInSlice(t *testing.T) {
+func TestIsStringInSlice(t *testing.T) {
 	cases := []struct {
 		needle   string
 		haystack []string
@@ -124,9 +124,9 @@ func TestStringInSlice(t *testing.T) {
 	}
 
 	for i, c := range cases {
-		got := StringInSlice(c.needle, c.haystack)
+		got := IsStringInSlice(c.needle, c.haystack)
 		if got != c.want {
-			t.Errorf("(%d) StringInSlice got: %t, want: %t", i, got, c.want)
+			t.Errorf("(%d) IsStringInSlice got: %t, want: %t", i, got, c.want)
 		}
 	}
 }